requirements.txt
Flask>=3.0.0
pandas>=2.1.0
nltk>=3.8.1
spacy>=3.7.0
scikit-learn>=1.3.0
joblib>=1.3.0
google-generativeai>=0.3.0
